refactor(rome_js_parser): Introduce `rome_parser` crate
