+++ crates/rome_css_syntax/src/lib.rs
@@ -52,6 +52,8 @@ impl CssSyntaxKind {
 }
 
 impl rome_rowan::SyntaxKind for CssSyntaxKind {
+    const TOMBSTONE: Self = CssSyntaxKind::TOMBSTONE;
+
     fn is_unknown(&self) -> bool {
         matches!(self, CSS_UNKNOWN)
     }
@@ -77,6 +79,10 @@ impl rome_rowan::SyntaxKind for CssSyntaxKind {
     #[inline]
     fn is_list(&self) -> bool {
         CssSyntaxKind::is_list(*self)
+    }
+
+    fn to_string(&self) -> Option<&'static str> {
+        CssSyntaxKind::to_string(self)
     }
 }
