+++ crates/rome_js_parser/Cargo.toml
@@ -14,6 +14,7 @@ rome_js_syntax = { path = "../rome_js_syntax" }
 rome_js_factory = { path = "../rome_js_factory" }
 rome_js_unicode_table = { path = "../rome_js_unicode_table" }
 rome_rowan = { path = "../rome_rowan" }
+rome_parser = { path = "../rome_parser" }
 drop_bomb = "0.1.5"
 bitflags = "1.3.2"
 indexmap = { workspace = true }
