+++ crates/rome_js_parser/src/lexer/buffered_lexer.rs
@@ -1,6 +1,6 @@
 use super::{LexContext, Lexer, LexerCheckpoint, ReLexContext, TextRange, TokenFlags};
-use crate::ParseDiagnostic;
 use rome_js_syntax::{JsSyntaxKind, JsSyntaxKind::EOF};
+use rome_parser::diagnostic::ParseDiagnostic;
 use std::collections::VecDeque;
 use std::iter::FusedIterator;
