+++ crates/rome_js_parser/src/lexer/mod.rs
@@ -30,13 +30,13 @@ pub use highlight::*;
 pub(crate) use buffered_lexer::BufferedLexer;
 pub use rome_js_syntax::*;
 
-use crate::ParseDiagnostic;
 use rome_diagnostics::location::FileId;
 use rome_js_syntax::JsSyntaxKind::*;
 use rome_js_unicode_table::{
     is_id_continue, is_id_start, lookup_byte,
     Dispatch::{self, *},
 };
+use rome_parser::diagnostic::ParseDiagnostic;
 
 use self::errors::invalid_digits_after_unicode_escape_sequence;
