+++ crates/rome_js_parser/src/parser/parsed_syntax.rs
@@ -1,7 +1,8 @@
 use crate::parser::parse_recovery::RecoveryResult;
 use crate::parser::ParsedSyntax::{Absent, Present};
-use crate::parser::{JsLanguageParser, LanguageParser, ParseRecovery, Parser, ToDiagnostic};
-use crate::{CompletedMarker, JsParser, Marker, ParseDiagnostic};
+use crate::parser::{JsLanguageParser, ParseRecovery};
+use crate::prelude::*;
+use crate::JsParser;
 use rome_rowan::TextRange;
 
 /// Syntax that is either present in the source tree or absent.
