+++ crates/rome_js_parser/src/prelude.rs
@@ -1,0 +1,2 @@
+pub(crate) use crate::parser::{JsParser, JsParserExtensions};
+pub use rome_parser::prelude::*;
