+++ crates/rome_parser/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "rome_parser"
+version = "0.0.0"
+edition = "2021"
+authors = ["Rome Tools Developers and Contributors"]
+description = "Shared infrastructure for Rome's parser"
+repository = "https://github.com/rome/tools"
+license = "MIT"
+
+
+[dependencies]
+rome_console = { path = "../rome_console" }
+rome_diagnostics = { path = "../rome_diagnostics" }
+rome_rowan = { path = "../rome_rowan" }
+drop_bomb = "0.1.5"
+
+#bitflags = "1.3.2"
+#indexmap = { workspace = true }
+#cfg-if = "1.0.0"
+#smallvec = { version = "1.8.0", features = ["union", "const_new"] }
+#tracing = { workspace = true }
