+++ crates/rome_parser/src/prelude.rs
@@ -1,0 +1,5 @@
+pub use crate::diagnostic::{ParseDiagnostic, ToDiagnostic};
+pub use crate::marker::{CompletedMarker, Marker};
+pub use crate::token_source::{BumpWithContext, NthToken, TokenSource};
+pub use crate::{token_set, TokenSet};
+pub use crate::{LanguageParser, Parser, Rewind};
